+++ framework-jmx-command-client/src/main/java/uk/gov/justice/framework/command/client/startup/Bootstrapper.java
@@ -2,6 +2,8 @@ package uk.gov.justice.framework.command.client.startup;
 
 import uk.gov.justice.framework.command.client.ReturnCode;
 
+import java.util.UUID;
+
 public class Bootstrapper {
 
     private final ObjectFactory objectFactory;
@@ -22,6 +24,7 @@ public class Bootstrapper {
     public static void main(String[] args) {
 
         final String command = "PING";
+        final String commandRuntimeId = "";
         final String contextName = "people";
 
         final String userName = "admin";
@@ -29,10 +32,11 @@ public class Bootstrapper {
 
 
         final String[] arguments = {
-                "-c", command,
-                "-u", userName,
-                "-pw", password,
-                "-cn", contextName
+            "-c", command,
+            "-rcid", commandRuntimeId,
+            "-u", userName,
+            "-pw", password,
+            "-cn", contextName
         };
 
         new Bootstrapper().startContainerAndRun(arguments);