+++ jmx/jmx-command-handling/src/main/java/uk/gov/justice/services/jmx/api/mbean/SystemCommander.java
@@ -1,6 +1,8 @@
 package uk.gov.justice.services.jmx.api.mbean;
 
 import static java.lang.String.format;
+import static java.util.Optional.empty;
+import static java.util.Optional.of;
 import static java.util.stream.Collectors.toList;
 
 import uk.gov.justice.services.jmx.api.CommandNotFoundException;
@@ -15,6 +17,7 @@ import uk.gov.justice.services.jmx.runner.AsynchronousCommandRunner;
 import uk.gov.justice.services.jmx.state.observers.SystemCommandStateBean;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import javax.inject.Inject;
@@ -39,30 +42,29 @@ public class SystemCommander implements SystemCommanderMBean {
     private CommandConverter commandConverter;
 
     @Inject
+    private SystemCommandVerifier systemCommandVerifier;
+
+    @Inject
     private Logger logger;
 
     @Override
     public UUID call(final String systemCommandName, final CommandRunMode commandRunMode) {
-
         logger.info(format("Received System Command '%s'", systemCommandName));
         logger.info(format("Running '%s' in '%s' mode", systemCommandName, commandRunMode));
 
-        final SystemCommand systemCommand = systemCommandLocator
-                .forName(systemCommandName)
-                .orElseThrow(() -> new UnrunnableSystemCommandException(format("The system command '%s' is not supported on this context.", systemCommandName)));
+        return doCall(systemCommandName, empty(), commandRunMode);
+    }
 
-        if (commandRunMode.isGuarded()) {
-            if (systemCommandStateBean.commandInProgress(systemCommand)) {
-                throw new UnrunnableSystemCommandException(format("Cannot run system command '%s'. A previous call to that command is still in progress.", systemCommand.getName()));
-            }
-        }
+    @Override
+    public UUID callWithRuntimeId(final String systemCommandName, final UUID commandRuntimeId, final CommandRunMode commandRunMode) {
+        logger.info(format("Received System Command '%s' with UUID '%s'", systemCommandName, commandRuntimeId));
+        logger.info(format("Running '%s' with UUID '%s' in '%s' mode", systemCommandName, commandRuntimeId, commandRunMode));
 
-        return asynchronousCommandRunner.run(systemCommand);
+        return doCall(systemCommandName, of(commandRuntimeId), commandRunMode);
     }
 
     @Override
     public List<SystemCommandDetails> listCommands() {
-
         return systemCommandScanner.findCommands().stream()
                 .map(commandConverter::toCommandDetails)
                 .collect(toList());
@@ -74,5 +76,21 @@ public class SystemCommander implements SystemCommanderMBean {
         return systemCommandStateBean
                 .getCommandStatus(commandId)
                 .orElseThrow(() -> new CommandNotFoundException(format("No SystemCommand found with id %s", commandId)));
+    }
+
+    private UUID doCall(final String systemCommandName, final Optional<UUID> commandRuntimeId, final CommandRunMode commandRunMode) {
+        final SystemCommand systemCommand = systemCommandLocator
+                .forName(systemCommandName)
+                .orElseThrow(() -> new UnrunnableSystemCommandException(format("The system command '%s' is not supported on this context.", systemCommandName)));
+
+        systemCommandVerifier.verify(systemCommand, commandRuntimeId);
+
+        if (commandRunMode.isGuarded()) {
+            if (systemCommandStateBean.commandInProgress(systemCommand)) {
+                throw new UnrunnableSystemCommandException(format("Cannot run system command '%s'. A previous call to that command is still in progress.", systemCommand.getName()));
+            }
+        }
+
+        return asynchronousCommandRunner.run(systemCommand, commandRuntimeId);
     }
 }