+++ jmx/jmx-command-handling/src/main/java/uk/gov/justice/services/jmx/command/HandlerMethodValidator.java
@@ -7,37 +7,49 @@ import uk.gov.justice.services.jmx.api.InvalidHandlerMethodException;
 import uk.gov.justice.services.jmx.api.command.SystemCommand;
 
 import java.lang.reflect.Method;
+import java.util.Optional;
 import java.util.UUID;
 
 public class HandlerMethodValidator {
 
 
-    public void checkHandlerMethodIsValid(final Method handlerMethod, final Object instance) {
+    public void checkHandlerMethodIsValid(final Method handlerMethod, final Object instance, final Optional<UUID> commandRuntimeId) throws InvalidHandlerMethodException {
 
         checkMethodPublic(handlerMethod, instance);
-        checkMethodParameter(handlerMethod, instance);
+        checkMethodParameters(handlerMethod, instance, commandRuntimeId);
     }
 
-    private void checkMethodPublic(final Method handlerMethod, final Object instance) {
+    private void checkMethodPublic(final Method handlerMethod, final Object instance) throws InvalidHandlerMethodException {
 
         if (!isPublic(handlerMethod.getModifiers())) {
             throw new InvalidHandlerMethodException(format("Handler method '%s' on class '%s' is not public.", handlerMethod.getName(), instance.getClass().getName()));
         }
     }
 
-    private void checkMethodParameter(final Method handlerMethod, final Object instance) {
+    private void checkMethodParameters(final Method handlerMethod, final Object instance, final Optional<UUID> commandRuntimeId) throws InvalidHandlerMethodException {
         final Class<?>[] parameterTypes = handlerMethod.getParameterTypes();
 
-        if (parameterTypes.length != 2) {
-            throw new InvalidHandlerMethodException(format("Invalid handler method '%s' on class '%s'. Method should have 2 parameters. First of type '%s' and second of type '%s'.",
-                    handlerMethod.getName(),
-                    instance.getClass().getName(),
-                    SystemCommand.class.getName(),
-                    UUID.class.getName()));
+        if (commandRuntimeId.isPresent()) {
+            if (parameterTypes.length != 3) {
+                throw new InvalidHandlerMethodException(format("Invalid handler method '%s' on class '%s'. Method should have 3 parameters. First of type '%s' and second of type '%s' and third of type '%s'.",
+                        handlerMethod.getName(),
+                        instance.getClass().getName(),
+                        SystemCommand.class.getName(),
+                        UUID.class.getName(),
+                        UUID.class.getName()));
+            }
+        } else {
+            if (parameterTypes.length != 2) {
+                throw new InvalidHandlerMethodException(format("Invalid handler method '%s' on class '%s'. Method should have 2 parameters. First of type '%s' and second of type '%s'.",
+                        handlerMethod.getName(),
+                        instance.getClass().getName(),
+                        SystemCommand.class.getName(),
+                        UUID.class.getName()));
+            }
         }
 
         if (!SystemCommand.class.isAssignableFrom(parameterTypes[0])) {
-            throw new InvalidHandlerMethodException(format("Invalid handler method '%s' on class '%s'. Method should have one parameter of type '%s'.",
+            throw new InvalidHandlerMethodException(format("Invalid handler method '%s' on class '%s'. Method should have first parameter of type '%s'.",
                     handlerMethod.getName(),
                     instance.getClass().getName(),
                     SystemCommand.class.getName()));