+++ jmx/jmx-command-handling/src/main/java/uk/gov/justice/services/jmx/command/SystemCommandHandlerProxy.java
@@ -2,11 +2,13 @@ package uk.gov.justice.services.jmx.command;
 
 import static java.lang.String.format;
 
+import uk.gov.justice.services.jmx.api.InvalidHandlerMethodException;
 import uk.gov.justice.services.jmx.api.SystemCommandInvocationException;
 import uk.gov.justice.services.jmx.api.command.SystemCommand;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.util.Optional;
 import java.util.UUID;
 
 public class SystemCommandHandlerProxy {
@@ -15,12 +17,18 @@ public class SystemCommandHandlerProxy {
     private final Method method;
     private final Object instance;
     private final HandlerMethodValidator handlerMethodValidator;
+    private final CommandHandlerMethodArgumentFactory commandHandlerMethodArgumentFactory;
 
-    public SystemCommandHandlerProxy(final String commandName, final Method method, final Object instance, final HandlerMethodValidator handlerMethodValidator) {
+    public SystemCommandHandlerProxy(
+            final String commandName, final Method method,
+            final Object instance,
+            final HandlerMethodValidator handlerMethodValidator,
+            final CommandHandlerMethodArgumentFactory commandHandlerMethodArgumentFactory) {
         this.commandName = commandName;
         this.method = method;
         this.instance = instance;
         this.handlerMethodValidator = handlerMethodValidator;
+        this.commandHandlerMethodArgumentFactory = commandHandlerMethodArgumentFactory;
     }
 
     public String getCommandName() {
@@ -31,12 +39,17 @@ public class SystemCommandHandlerProxy {
         return instance;
     }
 
-    public void invokeCommand(final SystemCommand systemCommand, final UUID commandId) throws SystemCommandInvocationException {
-
-        handlerMethodValidator.checkHandlerMethodIsValid(method, instance);
-
+    public void invokeCommand(final SystemCommand systemCommand, final UUID commandId, final Optional<UUID> commandRuntimeId) throws SystemCommandInvocationException {
         try {
-            method.invoke(instance, systemCommand, commandId);
+            handlerMethodValidator.checkHandlerMethodIsValid(method, instance, commandRuntimeId);
+            final Object[] methodArguments = commandHandlerMethodArgumentFactory.createMethodArguments(
+                    systemCommand,
+                    commandId,
+                    commandRuntimeId);
+            method.invoke(instance, methodArguments);
+        } catch (final InvalidHandlerMethodException e) {
+
+            throw new SystemCommandInvocationException(e.getMessage(), e);
         } catch (final IllegalAccessException e) {
 
             final String message = format("Failed to call method '%s()' on %s. Is the method public?",
@@ -57,5 +70,4 @@ public class SystemCommandHandlerProxy {
                     targetException);
         }
     }
-
 }