+++ jmx/jmx-command-handling/src/test/java/uk/gov/justice/services/jmx/command/TestCommandWithRuntimeId.java
@@ -2,16 +2,21 @@ package uk.gov.justice.services.jmx.command;
 
 import uk.gov.justice.services.jmx.api.command.BaseSystemCommand;
 
-public class TestCommand extends BaseSystemCommand {
+public class TestCommandWithRuntimeId extends BaseSystemCommand {
 
     public static final String TEST_COMMAND = "TEST_COMMAND";
 
-    public TestCommand() {
+    public TestCommandWithRuntimeId() {
         super(TEST_COMMAND, "This is a command used for testing");
     }
 
     @Override
+    public boolean requiresCommandRuntimeId() {
+        return true;
+    }
+
+    @Override
     public String commandRuntimeIdType() {
-        return "EVENT_ID";
+        return "eventId";
     }
 }