+++ jmx/jmx-command-handling/src/test/java/uk/gov/justice/services/jmx/runner/SystemCommandRunnerTest.java
@@ -1,29 +1,21 @@
 package uk.gov.justice.services.jmx.runner;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.nullValue;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.fail;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-import uk.gov.justice.services.framework.utilities.exceptions.StackTraceProvider;
-import uk.gov.justice.services.jmx.api.SystemCommandException;
-import uk.gov.justice.services.jmx.api.SystemCommandInvocationFailedException;
-import uk.gov.justice.services.jmx.command.SystemCommandHandlerProxy;
-import uk.gov.justice.services.jmx.command.SystemCommandStore;
-import uk.gov.justice.services.jmx.command.TestCommand;
-
-import java.util.UUID;
-
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.slf4j.Logger;
+import uk.gov.justice.services.jmx.command.SystemCommandHandlerProxy;
+import uk.gov.justice.services.jmx.command.SystemCommandStore;
+import uk.gov.justice.services.jmx.command.TestCommand;
+
+import java.util.Optional;
+import java.util.UUID;
+
+import static java.util.Optional.of;
+import static java.util.UUID.randomUUID;
+import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 public class SystemCommandRunnerTest {
@@ -32,50 +24,40 @@ public class SystemCommandRunnerTest {
     private SystemCommandStore systemCommandStore;
 
     @Mock
-    private StackTraceProvider stackTraceProvider;
-
-    @Mock
     private Logger logger;
 
     @InjectMocks
     private SystemCommandRunner systemCommandRunner;
 
     @Test
-    public void shouldFindTheCorrectProxyForTheCommandAndInvoke() throws Exception {
+    public void shouldFindTheCorrectProxyForTheCommandAndInvokeGivenCommandRuntimeId() throws Exception {
 
-        final UUID commandId = UUID.randomUUID();
+        final UUID commandId = randomUUID();
         final TestCommand testCommand = new TestCommand();
+        final Optional<UUID> commandRuntimeId = of(UUID.fromString("d3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"));
 
         final SystemCommandHandlerProxy systemCommandHandlerProxy = mock(SystemCommandHandlerProxy.class);
 
         when(systemCommandStore.findCommandProxy(testCommand)).thenReturn(systemCommandHandlerProxy);
 
-        systemCommandRunner.run(testCommand, commandId);
+        systemCommandRunner.run(testCommand, commandId, commandRuntimeId);
 
-        verify(systemCommandHandlerProxy).invokeCommand(testCommand, commandId);
+        verify(systemCommandHandlerProxy).invokeCommand(testCommand, commandId, commandRuntimeId);
+        verify(logger).info("Running system command 'TEST_COMMAND' with EVENT_ID 'd3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b'");
     }
 
     @Test
-    public void shouldThrowSystemCommandFailedExceptionIfCommandFails() throws Exception {
+    public void logGivenNoCommandRuntimeId() throws Exception {
 
-        final UUID commandId = UUID.randomUUID();
+        final UUID commandId = randomUUID();
         final TestCommand testCommand = new TestCommand();
-        final SystemCommandException systemCommandException = new SystemCommandException("Ooops");
-        final String stackTrace = "stack trace";
 
         final SystemCommandHandlerProxy systemCommandHandlerProxy = mock(SystemCommandHandlerProxy.class);
 
         when(systemCommandStore.findCommandProxy(testCommand)).thenReturn(systemCommandHandlerProxy);
-        doThrow(systemCommandException).when(systemCommandHandlerProxy).invokeCommand(testCommand, commandId);
-        when(stackTraceProvider.getStackTrace(systemCommandException)).thenReturn(stackTrace);
 
-        try {
-            systemCommandRunner.run(testCommand, commandId);
-            fail();
-        } catch (final SystemCommandInvocationFailedException expected) {
-            assertThat(expected.getMessage(), is("Failed to run System Command 'TEST_COMMAND'. Caused by uk.gov.justice.services.jmx.api.SystemCommandException: Ooops"));
-            assertThat(expected.getServerStackTrace(), is(stackTrace));
-            assertThat(expected.getCause(), is(nullValue()));
-        }
+        systemCommandRunner.run(testCommand, commandId, Optional.empty());
+
+        verify(logger).info("Running system command 'TEST_COMMAND'");
     }
 }